moved batch reducer to its own reducer
