+++ src/redux/reducers/assessment.reducer.js
@@ -25,19 +25,4 @@ const assessmentReducer = (state = [], action) => {
     }  
 }
 
-// Reducer will hold the batch id for students if active assessment is available
-const batch = (state = null, action) => {
-    switch(action.type) {
-        case 'SET_BATCH': 
-            return action.payload;
-        case 'EMPTY_BATCH':
-            return null
-        default:
-            return state;
-    }  
-}
-
-export default combineReducers({
-    assessmentReducer,
-    batch
-  });
+export default assessmentReducer;
